Move classes to payments section
