+++ docs/source/telegram.at-tree.rst
@@ -133,23 +133,12 @@ Available Types
     telegram.replykeyboardmarkup
     telegram.replykeyboardremove
     telegram.replyparameters
-    telegram.revenuewithdrawalstate
-    telegram.revenuewithdrawalstatefailed
-    telegram.revenuewithdrawalstatepending
-    telegram.revenuewithdrawalstatesucceeded
     telegram.sentwebappmessage
     telegram.shareduser
-    telegram.startransaction
-    telegram.startransactions
     telegram.story
     telegram.switchinlinequerychosenchat
     telegram.telegramobject
     telegram.textquote
-    telegram.transactionpartner
-    telegram.transactionpartnerfragment
-    telegram.transactionpartnerother
-    telegram.transactionpartnertelegramads
-    telegram.transactionpartneruser
     telegram.update
     telegram.user
     telegram.userchatboosts
