+++ telegram/_payment/stars.py
